Some refactoring and bug fixes
Still much to do.
Inventory desyncing when crafting will be fixed soon.
